Make editor buttons change based on editor state
Previously there was no indicator beyond the look of the editor itself
that any styles were active or not.

Now buttons that don't deal with text and highlight color become
'active' to show this. That 'active' style currently makes the text
of the button blue.

Dropdown lists now make the value known of the inline or block styles
instead of staying static like it previously was.

Text and highlight color button stylings will come later as there are
extra complications that the other buttons don't deal with.