Add color changes to text and highlight buttons
In order to facilitate the changes to the text color, changes had
to be made to the text color inline type. The highlight style didn't
have this problem.

A previously unknown issue was fixed, the editor's caret changed color
along with the text. So css was added to the top level to force the
caret to stay black to prevent it from turning functionally invisible
with white text.