Clean up Board Reducer and EditorButtonContainer
Made the Board Reducer less bulky by removing duplicate code into its
own function like how the Page Reducer works. This should make things
easier to deal with if functionality within the reducer needs to change.

The return statement of the EditorButtonContainer componenet easier
to read, and not quite as large. More changes need to be made to the
buttons and the container as a whole, but tidying the return should
help.