Vary editor button layouts for pages and notes
Made it so notes on the noteboard don't receive all editor buttons or
all funcionality that the editor has. A shorter list fits the bite
sized nature of the notes better and overall looks better and less
unwieldy.

Note: This is an edit of a previous commit that excluded an important
changed file and included one that didn't belong.
(bbecc00dff3e2928cfa8f2b580a854366e7c4d3c)