Fix issue NoteBoard wouldn't resize correctly
There was a problem where if you moved a note to the bottom right
corner of the NoteBoard to increase its size, only its width would
increase. This was due to the way the style object that controlled its
size was being updated. The function calls within NoteBoard used a
single snapshot of the style object, so all changes to the height were
gotten rid of while the width changed as the function to change the
width never recieves the updated style object.

Instead of updating in two separate functions, useEffect in NoteBoard
now facilitates these changes.