Add ability to change note colors
Only a short list of colors are added for now. If it's decided to add
more there is a system in place to do so.

Also fixed an issue with the note bottom positions and changing the
NoteBoard size. What numbers existed for the check stopped working
correctly, so they have been adjusted appropriately.