Make redux store track NoteBoard size
Reudx store and local storage now tracks the size of the NoteBoard,
also made it so if the size contained in state is less than the
inner window size, the display size now fills the screen. This fixes
a long standing issue caused by the component only checking the size
of the inner window on mount.