Move edit buttons on notes to not clip off screen
Not doing so causes issues where the editor buttons would create the
need to scroll over to see/use those off screen or be covered by the
main menu. Now when a note is close enough to an edge, once the drag
end action is executed, the note checks how close to the right or bottom
edge of the screen it is and moves the editor buttons accordingly.