Change Note Menu and lock effects on Notes
These changes make the lock effects clearer to the user. Previously
only the position lock changed anything about the note by getting rid
of the 'grab' cursor on the grab container. Now other changes are added
to the note and the note menu itself to better reflect what locks are
active without needing to open the 'Lock' submenu.