Add deletion confirmation and fixed note issues
Deletion now has a modal that appears, giving users an extra layer
of security to not accidentally delete notes.

Note editor button containers did not account for the scroll values
so the functionality that kept the buttons from overflowing off screen
didn't work if the noteboard was scrolled. New notes also always
appear at the top left of the screen, no matter how much scrolling has
been done. This makes it more convenient to add new notes if the board
is large.