Fix issue of note shrink when against board border
Before now, if you moved a note to the right border to increase
the board size, it would shrink the note to it's minimum width before
allowing you to move it further. This caused issues in increasing
the board size to fit the note's new location. Now when a note begins
it's drag state, its width is now locked until the drag event ends.

While this fixes the issue for the most part, there are times where
the note will shrink a tad on drag start. The cause of this issue is
currently unknown. This does not prevent the main functionality of this
commit's fix.