Add event listener cleanup and refactor dragEnd fn
Both event listeners in the PageContainer and EditorContainer were
not cleaned up on unmount, this has been fixed. The dragEnd function
was too long so it was refactored to help with readability and future
debugging if needed.

A typo was also fixed from the EditorButtonLayouts file and the one
file that used that layout.