Fix Note position awkwardness
Due to the way the note position was calculated in the drag events,
arbitrary values were needed to correctly evaluate container spacing.
This caused odd values to be used in checking and increasing the size
of the noteboard. This has been fixed.

The editor button class right value check was also fixed. There was
an oversight where the scrollX value was used twice, once in the wrong
spot in the class change breakpoint check.